move Utility.ts
