+++ frontend/src/Utility/Utility.ts
