+++ frontend/src/pages/AddFoodItem.tsx
@@ -17,7 +17,7 @@ import {DiaryEntry, FoodItem} from "../types/Diary.ts";
 import ModalAddFoodItem from "../components/modals/ModalAddFoodItem.tsx";
 import ModalFoodItems from "../components/modals/ModalFoodItems.tsx";
 import {v4 as uuidv4} from 'uuid';
-import {translateMealType} from "../Utility.ts";
+import {translateMealType} from "../Utility/Utility.ts";
 
 type AddFoodItemProps = {
     setCurrentRoute : (url:string) => void,
